+++ src/Shakebook.hs
@@ -1,5 +1,9 @@
 module Shakebook (
-  module Data.Aeson
+  module Control.Comonad.Cofree
+, module Control.Comonad.Store
+, module Control.Comonad.Store.Zipper
+, module Control.Comonad.Zipper.Extra
+, module Data.Aeson
 , module Development.Shake.Plus
 , module Shakebook.Conventions
 , module Shakebook.Data
@@ -7,8 +11,13 @@ module Shakebook (
 , module Shakebook.Feed
 , module Shakebook.Mustache
 , module Shakebook.Pandoc
+, module Text.Pandoc.Highlighting
 ) where
 
+import Control.Comonad.Cofree
+import Control.Comonad.Store
+import Control.Comonad.Store.Zipper
+import Control.Comonad.Zipper.Extra
 import Data.Aeson
 import Development.Shake.Plus
 import Shakebook.Conventions
@@ -17,3 +26,4 @@ import Shakebook.Defaults
 import Shakebook.Feed
 import Shakebook.Mustache
 import Shakebook.Pandoc
+import Text.Pandoc.Highlighting